Schemas Merge

+++ packages/express-backend/user-services.js
@@ -1,58 +1,67 @@
 import mongoose from "mongoose";
+import dotenv from 'dotenv';
 import userModel from "./user.js";
 
+dotenv.config();
 mongoose.set("debug", true);
 
-mongoose
-  .connect("mongodb://localhost:27017/users", {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  })
-  .catch((error) => console.log(error));
+mongoose.connect(process.env.MONGODB_URI, {
+    //useNewUrlParser: true,
+    //useUnifiedTopology: true
+}).catch((error) => console.log(error));
 
-function getUsers(name, job) {
-  let promise;
-  if (name === undefined && job === undefined) {
-    promise = userModel.find();
-  } else if (name && !job) {
-    promise = findUserByName(name);
-  } else if (job && !name) {
-    promise = findUserByJob(job);
-  }
-  return promise;
+function createUser(username) { // if we add a password, it should be hashed by this point
+    const userToAdd = new userModel({ username: username });
+    const promise = userToAdd.save();
+    return promise;
 }
 
-function findUserById(id) {
-  return userModel.findById(id);
+function setAsModerator(username) {
+    return userModel.findOne({ username: username }).updateOne({ type: 'moderator' });
 }
 
-function findUserbyNameAndJob(name, job) {
-  return userModel.find({ name: name, job: job });
+function setAsRegular(username) {
+    return userModel.findOne({ username: username }).updateOne({ type: 'regular' });
 }
 
-function addUser(user) {
-  const userToAdd = new userModel(user);
-  const promise = userToAdd.save();
-  return promise;
+function editUser(_id, username, about, profile) {
+    return userModel.findOne({ _id: _id }).updateOne({ username: username, about: about, profile: profile });
 }
 
-function findUserByName(name) {
-  return userModel.find({ name: name });
+function getUser(username) {
+    return userModel.findOne({ username: username });
 }
 
-function deleteByID(_id) {
-  return userModel.findByIdAndDelete(_id);
+function addFavorite(username, fileID, ...list) {
+    if (!list.includes(fileID)) {
+        listAdded = list.concat(fileID);
+        return userModel.findOne({ username: username }).updateOne({ favorites: listAdded });
+    }
+    return 0; // no action taken, item already in favorites
 }
 
-function findUserByJob(job) {
-  return userModel.find({ job: job });
+function removeFavorite(username, fileID, ...list) {
+    const index = list.indexOf(fileID);
+    if (index > -1) {
+        list.splice(index, 1); // id like to do this w/o modifying the original array
+        return userModel.findOne({ username: username }).updateOne({ favorites: list });
+    }
+    return 1;
+}
+
+function deleteMyUser(username) { // how to handle things made by removed users? can we null their data so theyre not technically removed from the db?
+    //log out user as well
+    return userModel.findOneAndDelete({ username: username });
+}
+
+function deleteOtherUser(_modId, _deleteId) { // need to add mod column in user table
+    const mod = userModel.findOne({ _id: _modId });
+    if (mod.type == 'moderator') {
+        return userModel.findByIdAndDelete(_id);
+    }
+    return 401; // Not authorized
 }
 
 export default {
-  addUser,
-  getUsers,
-  findUserById,
-  findUserByName,
-  findUserByJob,
-  deleteByID,
+    createUser, editUser, deleteMyUser, deleteOtherUser, getUser, addFavorite, removeFavorite, setAsModerator, setAsRegular
 };
